first

+++ EnvironmentalData.java
@@ -6,10 +6,10 @@ public abstract class EnvironmentalData implements Comparable<EnvironmentalData>
     private String locationName;      // Name of the geographical location
     private double latitude;           // Latitude of the location
     private double longitude;          // Longitude of the location
-    private int measurementTimestamp;  // Timestamp of the measurement
+    private long measurementTimestamp;  // Timestamp of the measurement
 
     // Constructor to initialize common attributes
-    public EnvironmentalData(String locationName, double latitude, double longitude, int measurementTimestamp) {
+    public EnvironmentalData(String locationName, double latitude, double longitude, long measurementTimestamp) {
         this.locationName = locationName;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -43,7 +43,7 @@ public abstract class EnvironmentalData implements Comparable<EnvironmentalData>
         this.longitude = longitude;
     }
 
-    public int getMeasurementTimestamp() {
+    public long getMeasurementTimestamp() {
         return measurementTimestamp;
     }
